docs(scripts): rename run_sample_sweep->run_agent_sweep; clarify CI=confidence interval; document manifest/compiler/harness (#133)
- rename run_sample_sweep.py -> run_agent_sweep.py (mine; updated the 2 importing
  scripts + doc). Left tested, not-mine legacy files (aggregate, aggregate_phase2,
  run_sweep, partition) for a maintainer-signed-off commit; README lists done-vs-pending.
- CI disambiguation: "CI = confidence interval, NOT continuous integration" in
  README + compare_harnesses docstring + output label.
- README: "How indexes reach the agent: RepoManifest <-> IndexCompiler <-> harness"
  — IndexCompiler builds + writes the manifest (AoT linking protocol); contexts load
  JIT (build_skill_contexts) or AoT (load_contexts_from_manifest); AgentRunner(manifest=)
  guards resources. Our cost harness BYPASSES the manifest (prebuilt.py stages /mnt/data,
  build_skill_contexts loads, no ResourceGuard). compile_table (routing) is orthogonal.
