+++ scripts/agent_compile/README.md
@@ -28,6 +28,39 @@ not a sweep):
 The **cost/ablation study** (graph-vs-retrieval, LocAgent) was added later and
 runs on a separate, simpler track (prebuilt indexes + reps), below.
 
+## How indexes reach the agent: RepoManifest ↔ IndexCompiler ↔ harness
+
+Three layers, and the cost-study harness deliberately uses only part of them:
+
+- **`IndexCompiler.compile_repo(repo, types)`** (`codeminer/compiler/index_compiler.py`)
+  builds the indexes (`bm25`, `vector`, `symbol_graph`) and writes a
+  **`RepoManifest`** (`repo_manifest.json`) — the *linking protocol* recording,
+  per index: `path`, `config`, `built_at`, `status` (fresh/failed), plus derived
+  capabilities. It's the source of truth a later run trusts instead of
+  re-deciding what exists.
+- **Getting indexes into agent skill `contexts`** has two routes:
+  - **JIT** — `build_skill_contexts(skill_ids, cache_dir, rebuild=…)`: infers the
+    index types the skills need (from each skill's `index_requirements`), builds
+    any missing via `IndexCompiler`, loads from the convention layout
+    `cache_dir/<type>`. No manifest needed.
+  - **AoT** — `load_contexts_from_manifest(manifest, skill_ids)`: loads index
+    paths/configs straight from a prebuilt manifest, validating freshness/status.
+- **At runtime**, `AgentRunner(manifest=…)` also uses the manifest for a
+  `ResourceGuard` preflight — skills whose indexes are missing/stale are excluded.
+
+**Where our harness sits:** the cost study **bypasses the manifest**.
+`prebuilt.py` symlinks the prebuilt `/mnt/data/codeminer` `vector` + `graph.pkl`
+into the `cache_dir/<type>` layout and builds BM25 fresh; then `run_agent_sweep`
+calls `build_skill_contexts(rebuild=False)` to *load* them via the JIT
+convention path. `AgentRunner` runs **without** a manifest (no ResourceGuard) and
+gets its skill subset directly via `allow_skills`. So `IndexCompiler` +
+`RepoManifest` are the **production AoT path** (Phase 3 "wire runtime"); the
+experiments ride the simpler convention loader because the indexes already exist.
+
+The **`compile_table` is orthogonal** to all of this: it maps *scenario → skill
+subset* (derived by `partition` + `aggregate_phase2`), i.e. *which skills to
+allow* — whereas the manifest is about *which indexes exist*.
+
 ## Scripts
 
 | file | purpose | track | status |
@@ -37,8 +70,8 @@ runs on a separate, simpler track (prebuilt indexes + reps), below.
 | `aggregate.py` | Phase-0 kill-switch report. | Phase 0 | **dead** (superseded) |
 | `aggregate_phase2.py` | A0–A6 aggregator: skill-invocation histogram, `compile_table` derivation, easy/hard saturation guard. | Phase 2 | active (sweep) |
 | `prebuilt.py` | stage the prebuilt `/mnt/data/codeminer` indexes (vector + graph, fresh BM25) into a cache. Shared by the cost track. | shared | active |
-| `run_sample_sweep.py` | **primary runner for the cost study**: one agent cell per (subset × instance × rep) on prebuilt indexes; honors the `cost_*` configs (incl. `default_tool_ids`, `system_prompt`). | cost/ablation | active |
-| `compare_harnesses.py` | compare two run dirs **accuracy-first**: mean accuracy + median-of-reps tokens, paired per-instance, with a Student-t 95% CI. *(was `aggregate_proof.py`)* | cost/ablation | active |
+| `run_agent_sweep.py` | **primary runner for the cost study**: one agent cell per (subset × instance × rep) on prebuilt indexes; honors the `cost_*` configs (incl. `default_tool_ids`, `system_prompt`). *(was `run_sample_sweep.py`)* | cost/ablation | active |
+| `compare_harnesses.py` | compare two run dirs **accuracy-first**: mean accuracy + median-of-reps tokens, paired per-instance, with a 95% **confidence interval** (Student-t). *(was `aggregate_proof.py`)* | cost/ablation | active |
 | `graph_recall_ablation.py` | retrieval-only (no LLM): does the call-graph recover a target file that **deep search misses**? Scans the dataset. | cost/ablation | active |
 | `graphrag_retrieve.py` | run the GraphRAG composer as a **standalone retrieval method** (files@k recall; `--no-graph` to ablate the graph). | cost/ablation | active |
 
@@ -56,10 +89,21 @@ runs on a separate, simpler track (prebuilt indexes + reps), below.
 | `cost_graphfriendly` | bm25 + embedding + callers/callees/trace | — | **stale** (never run) |
 | `phase0`, `sample` | A0–A6 full catalog | `sample*` | original Phase-0/2 sweep |
 
-## Proposed cleanup (not yet applied — needs sign-off; touches tests + #172)
+## "CI" disambiguation
 
-Intent-based renames for the dead/confusing files:
-`aggregate.py → DELETE` · `aggregate_phase2.py → aggregate_subset_sweep.py` ·
-`run_sweep.py → run_catalog_sweep.py` · `run_sample_sweep.py → run_agent_sweep.py`
-· delete `phase0.yaml`, `cost_graphfriendly.yaml`. These are cross-referenced by
-tests, so they go in a dedicated rename commit with reference updates.
+In this directory and in `docs/experiments/agent_compile.md`, **CI = confidence
+interval** (the 95% band on a mean token-delta), **not** continuous integration.
+
+## Naming cleanup — done + pending
+
+**Done** (files owned by this work, no test coupling):
+`aggregate_proof.py → compare_harnesses.py`, `run_sample_sweep.py →
+run_agent_sweep.py`.
+
+**Pending — needs maintainer sign-off** (these have their *own tests* and are
+not from this work, so they are NOT renamed/deleted here):
+`aggregate.py` (Phase-0, superseded — `test_aggregate.py`),
+`aggregate_phase2.py → aggregate_subset_sweep.py` (`test_aggregate_phase2.py`),
+`run_sweep.py → run_catalog_sweep.py`, `partition.py` (`test_partition.py`),
+and deleting the stale `phase0.yaml` / `cost_graphfriendly.yaml`. A dedicated
+commit with reference + test updates can land these once an owner confirms.
