+++ scripts/agent_compile/graphrag_retrieve.py
@@ -51,7 +51,7 @@ def _retrieve(instance_id: str, cfg: Any, prebuilt_dir: str, cache_root: str):
     from codeminer.agent.skills.registry import SkillRegistry
     from codeminer.eval.retrieval_eval import collect_targets
     from scripts.agent_compile.prebuilt import stage_prebuilt_indexes
-    from scripts.agent_compile.run_sample_sweep import (
+    from scripts.agent_compile.run_agent_sweep import (
         _load_dataset_rows,
         _load_full_contexts,
     )
@@ -80,7 +80,7 @@ def _retrieve(instance_id: str, cfg: Any, prebuilt_dir: str, cache_root: str):
 
 
 def main(argv: Optional[List[str]] = None) -> int:
-    from scripts.agent_compile.run_sample_sweep import SampleConfig
+    from scripts.agent_compile.run_agent_sweep import SampleConfig
 
     ap = argparse.ArgumentParser(description=__doc__)
     ap.add_argument("--instances-json", required=True, type=Path)
